Removed check for movetype in seismic sensor and made seismicsignature default to 0 for non ground units
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1679 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Units/UnitDef.h
@@ -79,7 +79,7 @@ struct UnitDef
 	int jammerRadius;
 	int sonarJamRadius;
 	int seismicRadius;
-	int seismicSignature;
+	float seismicSignature;
 	bool stealth;
 
 	float buildSpeed;
